clean: Factor ExecTool out of FileSystemTools
